Revert "refactor _schema.py into types/"
This reverts commit 51874971ab1a4debbe968f0ea75ea31374c2a778.
