+++ guidance/models/_azureai.py
@@ -1,7 +1,7 @@
 import logging
 from typing import TYPE_CHECKING, Any, Callable, ContextManager, Iterator, Optional, Union, cast
 
-from guidance.types import SamplingParams
+from guidance._schema import SamplingParams
 
 from .._ast import (
     JsonNode,
