+++ guidance/models/_openai.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from ..types import SamplingParams
+from .._schema import SamplingParams
 from ._base import Model
 from ._openai_base import (
     BaseOpenAIInterpreter,
