+++ guidance/models/experimental/_sglang.py
@@ -1,6 +1,6 @@
 from typing import Iterator, Optional
 
-from guidance.types import SamplingParams
+from guidance._schema import SamplingParams
 
 from ..._ast import GrammarNode, JsonNode, RegexNode, RuleNode
 from ...trace import OutputAttr, TextOutput
@@ -31,9 +31,9 @@ class SglangInterpreter(BaseOpenAIInterpreter):
 
         # Disable this check for now as all the supported endpoints have 'stop' support.
         if node.suffix:
-            raise ValueError("suffix not yet supported for sglang endpoint")
+            raise ValueError(f"suffix not yet supported for sglang endpoint")
         if node.stop_capture:
-            raise ValueError("stop_capture not yet supported for sglang endpoint")
+            raise ValueError(f"stop_capture not yet supported for sglang endpoint")
 
         kwargs = kwargs.copy()
         if node.temperature:
